+++ src/main/java/fr/estecka/variantscit/BakedModule.java
@@ -2,8 +2,8 @@ package fr.estecka.variantscit;
 
 import org.jetbrains.annotations.Nullable;
 import fr.estecka.variantscit.api.ICitModule;
-import net.minecraft.client.util.ModelIdentifier;
 import net.minecraft.item.ItemStack;
+import net.minecraft.util.Identifier;
 
 public record BakedModule(
 	VariantLibrary library,
@@ -11,7 +11,7 @@ public record BakedModule(
 )
 implements IItemModelProvider
 {
-	public @Nullable ModelIdentifier GetModelForItem(ItemStack stack){
+	public @Nullable Identifier GetModelForItem(ItemStack stack){
 		return logic.GetItemModel(stack, library);
 	}
 }